chore: remove `noir_wasm` dependency on `noirc_abi`
