+++ .github/workflows/publish-es-packages.yml
@@ -42,7 +42,6 @@ jobs:
             result/noirc_abi_wasm/web
 
   build-noir_wasm:
-    needs: [build-noirc_abi_wasm]
     runs-on: ubuntu-latest
     steps:
       - name: Checkout sources
@@ -58,14 +57,11 @@ jobs:
           key: noir-wasm
           save-if: false
 
-      - name: Download noirc_abi_wasm package artifact
-        uses: actions/download-artifact@v4
-        with:
-          name: noirc_abi_wasm
-          path: ./tooling/noirc_abi_wasm
-
       - name: Install Yarn dependencies
         uses: ./.github/actions/setup
+
+      - name: Build noir_js_types
+        run: yarn workspace @noir-lang/noir_js_types build
 
       - name: Build noir_wasm
         run: yarn workspace @noir-lang/noir_wasm build
