+++ compiler/wasm/package.json
@@ -80,6 +80,7 @@
     "webpack-cli": "^4.7.2"
   },
   "dependencies": {
+    "@noir-lang/types": "workspace:*",
     "pako": "^2.1.0"
   }
 }
